Add metadata about field types, ranges and whether all values are valid to the data returned in query results/tables. Automatically infer charting dimensions (possibly merging Year/Month/Day columns to dates). Generate chart of up to five measures against time dimension or row#.
Shelving this for now. There are a number of issues:

- if we have multiple data points for a single timestamp, we don't know whether
to sum, average, or incidence count. Currently we sum which is often wrong (e.g.
the gsod dataset which has temperature measurements from multiple weather stations).

- paginated table doesn't coexist nicely; we would need to either not paginate if
  we have a chart, or have a tabbed view so the user either sees a table or a chart
  but not both.

- categorical data is often encoded with integers, and we can't distinguish between
  these and valid quantitative fields

We could get around these to some extent with additional UI, but that is a slippery
slope, so instead the new plan is to use cell magics and the Google charting API.